+++ docs/include/samplecode/SpotifyClient/AddPlayerQueueItem.py
@@ -23,6 +23,8 @@ try:
     # add an item to the end of the user's current playback queue.
     trackUri:str = 'spotify:track:27JODWXo4VNa6s7HqDL9yQ'
     deviceId:str = None   # use currently playing device
+    #deviceId:str = "Web Player (Chrome)" # or device name
+    #deviceId:str = "0d1841b0976bae2a3a310dd74c0f3df354899bc8" # or device id
     print('\nAdding item to the users current playback queue:\n- "%s" ...' % (trackUri))
     spotify.AddPlayerQueueItem(trackUri, deviceId)