+++ docs/include/samplecode/SpotifyClient/PlayerMediaPlayTracks.py
@@ -22,6 +22,8 @@ try:
 
     # play single track on the specified Spotify Connect device.
     deviceId:str = None   # use currently playing device
+    #deviceId:str = "Web Player (Chrome)" # or device name
+    #deviceId:str = "0d1841b0976bae2a3a310dd74c0f3df354899bc8" # or device id
     uris:str='spotify:track:1kWUud3vY5ij5r62zxpTRy'  # Flawless
     print('\nPlaying media on Spotify Connect device: \nID: %s \n- %s' % (deviceId, uris.replace(',','\n- ')))
     spotify.PlayerMediaPlayTracks(uris, 0, deviceId)