+++ docs/include/samplecode/SpotifyClient/PlayerMediaSeek.py
@@ -23,6 +23,8 @@ try:
     # seek to the given position in the on the specified Spotify Connect device.
     positionMS:int = 25000
     deviceId:str = None   # use currently playing device
+    #deviceId:str = "Web Player (Chrome)" # or device name
+    #deviceId:str = "0d1841b0976bae2a3a310dd74c0f3df354899bc8" # or device id
     print('\nSeeking to %d milliseconds on Spotify Connect device:\n- "%s" ...' % (positionMS, deviceId))
     spotify.PlayerMediaSeek(positionMS, deviceId)