+++ docs/include/samplecode/SpotifyClient/PlayerMediaSkipNext.py
@@ -22,6 +22,8 @@ try:
 
     # skip to next track on the specified Spotify Connect device.
     deviceId:str = None   # use currently playing device
+    #deviceId:str = "Web Player (Chrome)" # or device name
+    #deviceId:str = "0d1841b0976bae2a3a310dd74c0f3df354899bc8" # or device id
     print('\nSkip to next track on Spotify Connect device:\n- "%s" ...' % str(deviceId))
     spotify.PlayerMediaSkipNext(deviceId)