+++ docs/include/samplecode/SpotifyClient/PlayerSetRepeatMode.py
@@ -22,6 +22,8 @@ try:
 
     # set repeat mode CONTEXT for the user's current playback device.
     deviceId:str = None   # use currently playing device
+    #deviceId:str = "Web Player (Chrome)" # or device name
+    #deviceId:str = "0d1841b0976bae2a3a310dd74c0f3df354899bc8" # or device id
     print('\nSet repeat mode CONTEXT for Spotify Connect device:\n- "%s" ...' % (str(deviceId)))
     spotify.PlayerSetRepeatMode('context', deviceId)