+++ docs/include/samplecode/SpotifyClient/PlayerSetVolume.py
@@ -18,6 +18,8 @@ try:
     # set the volume for the user's current playback device.
     volumePercent:int = 100
     deviceId:str = None   # use currently playing device
+    #deviceId:str = "Web Player (Chrome)" # or device name
+    #deviceId:str = "0d1841b0976bae2a3a310dd74c0f3df354899bc8" # or device id
     print('\nSet %d%% volume on Spotify Connect device:\n- "%s" ...' % (volumePercent, str(deviceId)))
     spotify.PlayerSetVolume(volumePercent, deviceId)